parse draw pixel parameters; closing #11

+++ src/display.h
@@ -28,7 +28,7 @@ SimpleTimer timer;
 void clear() { screen->clearScreen(); }
 
 uint16_t parseRGB(const String& rgbcode) {
-    uint16_t r, g, b;
+    uint32_t r, g, b;
     sscanf(rgbcode.c_str(), "#%02x%02x%02x", &r, &g, &b);
     return screen->color565(r, g, b);
 }
